Obem Navigation Bar Complete

+++ database/migrations/2022_06_11_135215_create_obem_article_media_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('mime_type');
             $table->string('media_file_name');
             $table->string('media_file_path')->unique();
-            $table->integer('article_id');
+            $table->bigInteger('article_id');
             $table->timestamps();
         });
     }
